+++ examples/Ex106_PSC_SurfaceRecombination.jl
@@ -133,9 +133,9 @@ function main(;
     ## Possible choices: Stationary, Transient
     data.modelType = Transient
 
-    ## Possible choices: Boltzmann, FermiDiracOneHalfBednarczyk, FermiDiracOneHalfTeSCA,
-    ## FermiDiracMinusOne, Blakemore
-    data.F = [FermiDiracOneHalfTeSCA, FermiDiracOneHalfTeSCA, FermiDiracMinusOne]
+    ## The default for electrons and holes is Boltzmann. Here, we set it to a more general statistics function
+    data.F[p.iphin] = FermiDiracOneHalfTeSCA
+    data.F[p.iphip] = FermiDiracOneHalfTeSCA
 
     data.bulkRecombination = set_bulk_recombination(;
         iphin = p.iphin, iphip = p.iphip,
@@ -152,11 +152,8 @@ function main(;
     data.boundaryType[p.bregionDonor] = OhmicContact
 
     ## Present ionic vacancies in perovskite layer
+    ## by default the statistics function is set to FermiDiracMinusOne to limit ion depletion
     enable_ionic_carrier!(data, ionicCarrier = p.iphia, regions = [p.regionIntrinsic])
-
-    ## Choose flux discretization scheme: ScharfetterGummel, ScharfetterGummelGraded,
-    ## ExcessChemicalPotential, ExcessChemicalPotentialGraded, DiffusionEnhanced, GeneralizedSG
-    data.fluxApproximation .= ExcessChemicalPotential
 
     if test == false
         println("*** done\n")
@@ -240,15 +237,33 @@ function main(;
 
     end # time loop
 
+    solution = sol.u[end]
+
     biasValues = contactVoltageFunction[p.bregionAcceptor].(tvalues)
 
     if plotting
+        label_solution, label_density, label_energy, label_BEE = set_plotting_labels(data)
+        ## add labels for anion vacancy
+        label_energy[1, p.iphia] = "\$E_a-q\\psi\$"; label_energy[2, p.iphia] = "\$ - q \\varphi_a\$"; label_BEE[p.iphia] = "\$E_a\$"
+        label_density[p.iphia] = "\$ n_a \$";      label_solution[p.iphia] = "\$ \\varphi_a\$"
+
+        Plotter.figure()
+        plot_densities(Plotter, ctsys, solution, "bias \$\\Delta u\$ ", label_density)
+        Plotter.figure()
+        plot_solution(Plotter, ctsys, solution, "bias \$\\Delta u\$", label_solution)
+
         Plotter.figure()
         Plotter.plot(tvalues, biasValues, marker = "o")
         Plotter.xlabel("time [s]")
         Plotter.ylabel("bias [V]")
+        Plotter.tight_layout()
         Plotter.figure()
-        plot_IV(Plotter, biasValues[2:end], IV, "Total current")
+        ###########
+        Plotter.plot(biasValues[2:end], IV .* (cm^2) .* 1.0e3, linewidth = 5, color = "darkblue")
+        Plotter.grid()
+        Plotter.xlabel("bias [V]")
+        Plotter.ylabel("current density [mAcm\$^{-2} \$]")
+        Plotter.tight_layout()
     end
 
     if test == false
@@ -275,7 +290,7 @@ end # main
 
 function test()
     testval = -0.5965444263524541; testvalvacancyEnergyCalculation = -0.5966729068541846
-    return main(test = true) ≈ testval && main(test = true, vacancyEnergyCalculation = false) ≈ testvalvacancyEnergyCalculation
+    return main(test = true, vacancyEnergyCalculation = true) ≈ testval && main(test = true, vacancyEnergyCalculation = false) ≈ testvalvacancyEnergyCalculation
 end