+++ examples/Ex202_Laser_simple.jl
@@ -135,9 +135,7 @@ function main(;
     ## Possible choices: Stationary, Transient
     data.modelType = Stationary
 
-    ## Possible choices: Boltzmann, FermiDiracOneHalfBednarczyk, FermiDiracOneHalfTeSCA,
-    ## FermiDiracMinusOne, Blakemore.
-    ## Can be a vector with different statistics (for n and p).
+    ## The default for electrons and holes is Boltzmann. Here, we set it to a more general statistics function
     data.F .= FermiDiracOneHalfTeSCA
 
     data.bulkRecombination = set_bulk_recombination(;
@@ -151,8 +149,6 @@ function main(;
     data.boundaryType[p.bregionDonor1] = OhmicContact     # bottom boundary Dirichlet condition
     #                                                     # rest is set to Neumann by default
 
-    data.fluxApproximation .= ExcessChemicalPotential
-    #data.fluxApproximation .= ScharfetterGummel          # if F=Boltzmann
 
     if test == false
         println("*** done\n")